+++ packages/rax-miniapp-runtime-webpack-plugin/CHANGELOG.md
@@ -1,11 +1,16 @@
 # Changelog
 
-## [5.0.5] - 2023-08-23
+## [5.0.6] Beta
 
 ### Added
 
+- [5.0.5-0] Support `page-container` native component in wechat-miniprogram
+
+## [5.0.5] - 2023-08-23
+
+### Fixed
+
 - Support onChooseAvatar event in buton in wechat
-- Support `page-container` native component in wechat-miniprogram
 
 ## [5.0.4] - 2022-08-30